ChallengeLength must be treated different depending on extended security on/off:
ChallengeLength (1 byte): When the CAP_EXTENDED_SECURITY bit is set, the server MUST set this value to zero and clients MUST ignore this value.

and when CAP_EXTENDED_SECURITY is not set:
ChallengeLength (1 byte): This field MUST be 0x00 or 0x08. The length of the random challenge used in challenge/response authentication. If the server does not support challenge/response authentication, this field MUST be 0x00. This field is often referred to in older documentation as EncryptionKeyLength.




git-svn-id: http://impacket.googlecode.com/svn/trunk@711 ad9a8d78-5211-11de-8950-a5d9c972bc64